+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `sandbox.Geometry.from_stl`. It reads a `stl` file and is able to compute what mesh points are inside or outside the geometry, so we can specify the geometry for a very customized immersed boundary method. By [@fschuch](https://github.com/fschuch) and [@nbeb](https://github.com/nbeb).
 
+### Fixed
+
+- `fix_bug` at [gene_epsi_3D](xcompact3d-toolbox/genepsi.py) was not working properly (#3), by [@fschuch](https://github.com/fschuch).
+
 ## [1.0.1] - 2021-09-23
 
 ### Modified